moved home to components and arrow func implimented
