+++ src/__test__/Api.test.js
@@ -1,0 +1,11 @@
+import api from '../API/api';
+
+describe('api API', () => {
+  it('Api testing', () => {
+  
+    then(response => {
+      expect(response).toEqual('hello');
+    })
+
+  });
+});
