+++ src/components/Header.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import '../styles/Header.css';
 
-function Header() {
+const Header = () => {
   return (
     <div className="navbar">
       <h1 className="left-head">Crypto</h1>
